requirements.txt

python-dotenv
bcrypt
email-validator
Flask
Flask-CORS
Flask-SQLAlchemy
psycopg2-binary
PyJWT
python-dotenv
python-multipart
Flask-JWT-Extended